moved headerdropdown into the header instead of the userlogo and cleaned up both the userlogo and header dropdown components
